% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R,
%   R/MsBackendRawFileReader.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{[,MsBackendRawFileReader-method}
\alias{filterScan}
\alias{scanType}
\alias{backendInitialize,MsBackendRawFileReader-method}
\alias{show,MsBackendRawFileReader-method}
\alias{peaksData,MsBackendRawFileReader-method}
\alias{[,MsBackendRawFileReader,ANY,ANY-method}
\alias{intensity,MsBackendRawFileReader-method}
\title{Internal page for hidden aliases}
\usage{
filterScan(object, filter, ...)

scanType(object, ...)

\S4method{backendInitialize}{MsBackendRawFileReader}(object, files, ..., BPPARAM = bpparam())

\S4method{show}{MsBackendRawFileReader}(object)

\S4method{peaksData}{MsBackendRawFileReader}(object, ..., BPPARAM = bpparam())

\S4method{[}{MsBackendRawFileReader,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{intensity}{MsBackendRawFileReader}(object, ..., BPPARAM = bpparam())
}
\value{
Not applicable
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\examples{
beRaw <- Spectra::backendInitialize(MsBackendRawFileReader::MsBackendRawFileReader(),
  files = rawrr::sampleFilePath())
beRaw
}
\keyword{internal}
