% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendRawFileReader.R
\docType{class}
\name{MsBackendRawFileReader-class}
\alias{MsBackendRawFileReader-class}
\alias{MsBackendRawFileReader}
\title{RawFileReader-based backend for Spectra}
\description{
The `MsBackendRawFileReader` inherits all slots and methods from the base
`MsBackendDataFrame` (in-memory) backend. It overrides the base `mz` and
`intensity` methods as well as `peaks` to read the respective data from
the original raw data files.


The `backendInitialize` method reads the header data from the raw files and
hence fills the `spectraData` slot. Note that this method could be called
several times, e.g. also to *re-fill* `spectraData` after dropping some of
its columns.
}
\references{
\doi{10.1021/acs.jproteome.8b00173}
}
\author{
Christian Panse <cp@fgcz.ethz.ch>, 2019-06-15 
adapted from the \code{\link[Spectra]{MsBackendMzR}.R} file by Johannes Rainer
}
