% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/MsBackendRawFileReader.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{[,MsBackendRawFileReader-method}
\alias{backendInitialize,MsBackendRawFileReader-method}
\alias{show,MsBackendRawFileReader-method}
\alias{intensity,MsBackendRawFileReader-method}
\alias{mz,MsBackendRawFileReader-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{backendInitialize}{MsBackendRawFileReader}(object, files, ...,
  BPPARAM = bpparam())

\S4method{show}{MsBackendRawFileReader}(object)

\S4method{intensity}{MsBackendRawFileReader}(object)

\S4method{mz}{MsBackendRawFileReader}(object)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
